Resolve Java build issue
